Fixed some bugs
git-svn-id: https://fusion.wpi.edu/svn/rbe3001a15team2/trunk@55 015edbb1-928d-476b-afdc-d2170ecc5f68

+++ RBELib/include/RBELib/pot.h
@@ -23,7 +23,7 @@ typedef struct
 {
 	long min;
 	int adcChannel;
-	int scaler;
+	double scaler;
 } potentiometer;
 
 potentiometer pots[2];
