+++ RBELib/src/pot.c
@@ -11,7 +11,7 @@ void initPot(int pot, int channel, potCalibration calib)
 {
 	pots[pot].adcChannel = channel;
 	pots[pot].min = calib.DEG_0;
-	pots[pot].scaler = (int)(180.0 / (calib.DEG_180 - calib.DEG_0));
+	pots[pot].scaler = 180.0 / (calib.DEG_180 - calib.DEG_0);
 }
 
 int potAngle(int pot)
